comminize trigger new and deprec. cache
