+++ core/src/main/java/apoc/CoreApocGlobalComponents.java
@@ -1,8 +1,7 @@
 package apoc;
 
 import apoc.cypher.CypherInitializer;
-import apoc.trigger.TriggerDeprecatedProcsHandler;
-import apoc.trigger.TriggerHandlerRead;
+import apoc.trigger.TriggerHandler;
 import org.neo4j.annotations.service.ServiceProvider;
 import org.neo4j.kernel.availability.AvailabilityListener;
 import org.neo4j.kernel.internal.GraphDatabaseAPI;
@@ -10,7 +9,6 @@ import org.neo4j.kernel.lifecycle.Lifecycle;
 
 import java.util.Collection;
 import java.util.Collections;
-import java.util.List;
 import java.util.Map;
 
 @ServiceProvider
@@ -18,25 +16,18 @@ public class CoreApocGlobalComponents implements ApocGlobalComponents {
 
     @Override
     public Map<String,Lifecycle> getServices(GraphDatabaseAPI db, ApocExtensionFactory.Dependencies dependencies) {
-        return Map.of("triggerDeprecated", 
-                new TriggerDeprecatedProcsHandler(db, 
-                        dependencies.databaseManagementService(),
-                        dependencies.apocConfig(),
-                        dependencies.log().getUserLog(TriggerDeprecatedProcsHandler.class),
-                        dependencies.pools(),
-                        dependencies.scheduler()),
-                "trigger", 
-                new TriggerHandlerRead(db, 
-                        dependencies.databaseManagementService(),
-                        dependencies.log().getUserLog(TriggerHandlerRead.class),
-                        dependencies.pools(),
-                        dependencies.scheduler())
+        return Collections.singletonMap("trigger", new TriggerHandler(db,
+                dependencies.databaseManagementService(),
+                dependencies.apocConfig(),
+                dependencies.log().getUserLog(TriggerHandler.class),
+                dependencies.pools(),
+                dependencies.scheduler())
         );
     }
 
     @Override
     public Collection<Class> getContextClasses() {
-        return List.of(TriggerDeprecatedProcsHandler.class, TriggerHandlerRead.class);
+        return Collections.singleton(TriggerHandler.class);
     }
 
     @Override
