+++ core/src/main/java/apoc/trigger/TriggerHandler.java
@@ -1,5 +1,6 @@
 package apoc.trigger;
 
+import apoc.ApocConfig;
 import apoc.Pools;
 import apoc.SystemLabels;
 import apoc.SystemPropertyKeys;
@@ -13,37 +14,37 @@ import org.neo4j.graphdb.event.TransactionData;
 import org.neo4j.graphdb.event.TransactionEventListener;
 import org.neo4j.internal.helpers.collection.Iterators;
 import org.neo4j.internal.helpers.collection.MapUtil;
+import org.neo4j.internal.helpers.collection.Pair;
 import org.neo4j.kernel.lifecycle.LifecycleAdapter;
 import org.neo4j.logging.Log;
 import org.neo4j.scheduler.Group;
 import org.neo4j.scheduler.JobHandle;
 import org.neo4j.scheduler.JobScheduler;
 
+import java.util.Collections;
 import java.util.LinkedHashMap;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.function.Function;
+import java.util.stream.Collectors;
 
+import static apoc.ApocConfig.APOC_TRIGGER_ENABLED;
 import static apoc.ApocConfig.apocConfig;
-import static apoc.trigger.TriggerHandlerWrite.getTriggerNodes;
-import static apoc.trigger.TriggerHandlerWrite.withSystemDb;
 
-
-public class TriggerHandlerRead extends LifecycleAdapter implements TransactionEventListener<Void> {
-
-    public static final int TRIGGER_RELOAD_DEFAULT_VALUE = 3000;
-    public static final String TRIGGER_RELOAD = "apoc.trigger.reload";
+public class TriggerHandler extends LifecycleAdapter implements TransactionEventListener<Void> {
 
     private enum Phase {before, after, rollback, afterAsync}
-    
-    private final ConcurrentHashMap<String, Map<String,Object>> activeTriggers = new ConcurrentHashMap<>();
 
+    public static final String TRIGGER_REFRESH = "apoc.trigger.refresh";
+
+    private final ConcurrentHashMap<String, Map<String,Object>> activeTriggers = new ConcurrentHashMap();
     private final Log log;
     private final GraphDatabaseService db;
     private final DatabaseManagementService databaseManagementService;
+    private final ApocConfig apocConfig;
     private final Pools pools;
-    
     private final JobScheduler jobScheduler;
 
     private long lastUpdate;
@@ -52,20 +53,52 @@ public class TriggerHandlerRead extends LifecycleAdapter implements TransactionE
 
     private final AtomicBoolean registeredWithKernel = new AtomicBoolean(false);
 
-    public TriggerHandlerRead(GraphDatabaseService db, DatabaseManagementService databaseManagementService,
-                              Log log, Pools pools, JobScheduler jobScheduler) {
+    public static final String NOT_ENABLED_ERROR = "Triggers have not been enabled." +
+            " Set 'apoc.trigger.enabled=true' in your apoc.conf file located in the $NEO4J_HOME/conf/ directory.";
+
+    public TriggerHandler(GraphDatabaseService db, DatabaseManagementService databaseManagementService,
+                          ApocConfig apocConfig, Log log,
+                          Pools pools, JobScheduler jobScheduler) {
         this.db = db;
         this.databaseManagementService = databaseManagementService;
+        this.apocConfig = apocConfig;
         this.log = log;
         this.pools = pools;
         this.jobScheduler = jobScheduler;
     }
 
+    private boolean isEnabled() {
+        return apocConfig.getBoolean(APOC_TRIGGER_ENABLED);
+    }
 
-    public void reconcileKernelRegistration() {
-        final boolean hasTriggers = withSystemDb(tx -> getTriggerNodes(db.databaseName(), tx)
-                .hasNext());
-        reconcileKernelRegistration(hasTriggers);
+    public void checkEnabled() {
+        if (!isEnabled()) {
+            throw new RuntimeException(NOT_ENABLED_ERROR);
+        }
+    }
+
+    private void updateCache() {
+        activeTriggers.clear();
+
+        lastUpdate = System.currentTimeMillis();
+
+        withSystemDb(tx -> {
+            tx.findNodes(SystemLabels.ApocTrigger,
+                    SystemPropertyKeys.database.name(), db.databaseName()).forEachRemaining(
+                    node -> activeTriggers.put(
+                            (String) node.getProperty(SystemPropertyKeys.name.name()),
+                            MapUtil.map(
+                                    "statement", node.getProperty(SystemPropertyKeys.statement.name()),
+                                    "selector", Util.fromJson((String) node.getProperty(SystemPropertyKeys.selector.name()), Map.class),
+                                    "params", Util.fromJson((String) node.getProperty(SystemPropertyKeys.params.name()), Map.class),
+                                    "paused", node.getProperty(SystemPropertyKeys.paused.name())
+                            )
+                    )
+            );
+            return null;
+        });
+
+        reconcileKernelRegistration();
     }
 
     /**
@@ -75,9 +108,9 @@ public class TriggerHandlerRead extends LifecycleAdapter implements TransactionE
      * For most deployments this isn't an issue, since you can turn the config flag off, but in large fleet deployments
      * it's nice to have uniform config, and then the memory savings on databases that don't use triggers is good.
      */
-    public synchronized void reconcileKernelRegistration(boolean hasTriggers) {
+    private synchronized void reconcileKernelRegistration() {
         // Register if there are triggers
-        if (hasTriggers) {
+        if (activeTriggers.size() > 0) {
             // This gets called every time triggers update; only register if we aren't already
             if(registeredWithKernel.compareAndSet(false, true)) {
                 databaseManagementService.registerTransactionEventListener(db.databaseName(), this);
@@ -90,8 +123,77 @@ public class TriggerHandlerRead extends LifecycleAdapter implements TransactionE
         }
     }
 
+    public Map<String, Object> add(String name, String statement, Map<String,Object> selector) {
+        return add(name, statement, selector, Collections.emptyMap());
+    }
+
+    public Map<String, Object> add(String name, String statement, Map<String,Object> selector, Map<String,Object> params) {
+        checkEnabled();
+        Map<String, Object> previous = activeTriggers.get(name);
+
+        withSystemDb(tx -> {
+            Node node = Util.mergeNode(tx, SystemLabels.ApocTrigger, null,
+                    Pair.of(SystemPropertyKeys.database.name(), db.databaseName()),
+                    Pair.of(SystemPropertyKeys.name.name(), name));
+            node.setProperty(SystemPropertyKeys.statement.name(), statement);
+            node.setProperty(SystemPropertyKeys.selector.name(), Util.toJson(selector));
+            node.setProperty(SystemPropertyKeys.params.name(), Util.toJson(params));
+            node.setProperty(SystemPropertyKeys.paused.name(), false);
+            setLastUpdate(tx);
+            return null;
+        });
+
+        updateCache();
+        return previous;
+    }
+
+    public Map<String, Object> remove(String name) {
+        checkEnabled();
+        Map<String, Object> previous = activeTriggers.remove(name);
+
+        withSystemDb(tx -> {
+            tx.findNodes(SystemLabels.ApocTrigger,
+                            SystemPropertyKeys.database.name(), db.databaseName(),
+                            SystemPropertyKeys.name.name(), name)
+                    .forEachRemaining(node -> node.delete());
+            setLastUpdate(tx);
+            return null;
+        });
+        updateCache();
+        return previous;
+    }
+
+    public Map<String, Object> updatePaused(String name, boolean paused) {
+        checkEnabled();
+        withSystemDb(tx -> {
+            tx.findNodes(SystemLabels.ApocTrigger,
+                            SystemPropertyKeys.database.name(), db.databaseName(),
+                            SystemPropertyKeys.name.name(), name)
+                    .forEachRemaining(node -> node.setProperty(SystemPropertyKeys.paused.name(), paused));
+            setLastUpdate(tx);
+            return null;
+        });
+        updateCache();
+        return activeTriggers.get(name);
+    }
+
+    public Map<String, Object> removeAll() {
+        checkEnabled();
+        Map<String, Object> previous = activeTriggers
+                .entrySet().stream().collect(Collectors.toMap(e -> e.getKey(), e -> e.getValue()));
+        withSystemDb(tx -> {
+            tx.findNodes(SystemLabels.ApocTrigger,
+                            SystemPropertyKeys.database.name(), db.databaseName() )
+                    .forEachRemaining(node -> node.delete());
+            setLastUpdate(tx);
+            return null;
+        });
+        updateCache();
+        return previous;
+    }
+
     public Map<String,Map<String,Object>> list() {
-        TriggerHandlerWrite.checkEnabled();
+        checkEnabled();
         return Map.copyOf(activeTriggers);
     }
 
@@ -164,7 +266,7 @@ public class TriggerHandlerRead extends LifecycleAdapter implements TransactionE
             }
         });
         if (!exceptions.isEmpty()) {
-            throw new RuntimeException("Error executing triggers "+ exceptions);
+            throw new RuntimeException("Error executing triggers "+exceptions.toString());
         }
     }
 
@@ -176,43 +278,12 @@ public class TriggerHandlerRead extends LifecycleAdapter implements TransactionE
     @Override
     public void start() throws Exception {
         updateCache();
-        long refreshInterval = apocConfig().getInt(TRIGGER_RELOAD, TRIGGER_RELOAD_DEFAULT_VALUE);
+        long refreshInterval = apocConfig().getInt(TRIGGER_REFRESH, 60000);
         restoreTriggerHandler = jobScheduler.scheduleRecurring(Group.STORAGE_MAINTENANCE, () -> {
             if (getLastUpdate() > lastUpdate) {
                 updateCache();
             }
         }, refreshInterval, refreshInterval, TimeUnit.MILLISECONDS);
-    }
-
-    private long getLastUpdate() {
-        return withSystemDb(tx -> {
-            Node node = tx.findNode(SystemLabels.ApocTriggerMeta, SystemPropertyKeys.database.name(), db.databaseName());
-            return node == null ? 0L : (long) node.getProperty(SystemPropertyKeys.lastUpdated.name());
-        });
-    }
-    
-    private void updateCache() {
-        activeTriggers.clear();
-
-        lastUpdate = System.currentTimeMillis();
-
-        withSystemDb(tx -> {
-            tx.findNodes(SystemLabels.ApocTrigger,
-                    SystemPropertyKeys.database.name(), db.databaseName()).forEachRemaining(
-                    node -> activeTriggers.put(
-                            (String) node.getProperty(SystemPropertyKeys.name.name()),
-                            MapUtil.map(
-                                    "statement", node.getProperty(SystemPropertyKeys.statement.name()),
-                                    "selector", Util.fromJson((String) node.getProperty(SystemPropertyKeys.selector.name()), Map.class),
-                                    "params", Util.fromJson((String) node.getProperty(SystemPropertyKeys.params.name()), Map.class),
-                                    "paused", node.getProperty(SystemPropertyKeys.paused.name())
-                            )
-                    )
-            );
-            return null;
-        });
-
-        reconcileKernelRegistration();
     }
 
     @Override
@@ -223,6 +294,30 @@ public class TriggerHandlerRead extends LifecycleAdapter implements TransactionE
         if (restoreTriggerHandler != null) {
             restoreTriggerHandler.cancel();
         }
+    }
+
+    private <T> T withSystemDb(Function<Transaction, T> action) {
+        try (Transaction tx = apocConfig.getSystemDb().beginTx()) {
+            T result = action.apply(tx);
+            tx.commit();
+            return result;
+        }
+    }
+
+    private long getLastUpdate() {
+        return withSystemDb( tx -> {
+            Node node = tx.findNode(SystemLabels.ApocTriggerMeta, SystemPropertyKeys.database.name(), db.databaseName());
+            return node == null ? 0L : (long) node.getProperty(SystemPropertyKeys.lastUpdated.name());
+        });
+    }
+
+    private void setLastUpdate(Transaction tx) {
+        Node node = tx.findNode(SystemLabels.ApocTriggerMeta, SystemPropertyKeys.database.name(), db.databaseName());
+        if (node == null) {
+            node = tx.createNode(SystemLabels.ApocTriggerMeta);
+            node.setProperty(SystemPropertyKeys.database.name(), db.databaseName());
+        }
+        node.setProperty(SystemPropertyKeys.lastUpdated.name(), System.currentTimeMillis());
     }
 
 }
